refactor(rest): extract external-source-node routes into routes/source_nodes.py
Moves GET/POST /api/scopes/{scope}/source-nodes -- the external-model
change-tracking group -- out of the create_app closure into a standalone
APIRouter. This group needed no RestContext field: the DB pool comes off
request.app.state the same way every other still-in-closure DB route reads
it, and none of its helpers (_source_nodes_pool, _source_node_json,
_parse_source_node, _SOURCE_NODE_WRITE_LIMIT) had another caller, so they
moved in full with no aliases left behind.

Drops the now-unused top-level `import datetime` from app.py (its only
module-scope use was in the extracted routes); every remaining `datetime`
reference in the closure already shadows it locally with its own `from
datetime import datetime, timezone`, which also clears a cascade of
ruff F811 redefinition warnings those locals were tripping now that the
outer import was dead.
